+++ package.json
@@ -47,6 +47,7 @@
     "@playwright/test": "^1.40.1",
     "@types/events": "^3.0.3",
     "@types/node": "^20.10.3",
+    "@types/uuid": "^9.0.8",
     "@typescript-eslint/eslint-plugin": "^6.13.2",
     "eslint": "^8.55.0",
     "eslint-config-prettier": "^9.1.0",
@@ -62,10 +63,10 @@
     "typescript": "^4.9.5"
   },
   "dependencies": {
-    "@jellyfish-dev/membrane-webrtc-js": "^0.7.1",
     "events": "^3.3.0",
     "ts-proto": "^1.165.0",
-    "typed-emitter": "^2.1.0"
+    "typed-emitter": "^2.1.0",
+    "uuid": "^9.0.1"
   },
   "directories": {
     "example": "examples"
